+++ eric-sdk/tests/test_send.rs
@@ -1,5 +1,5 @@
 use anyhow::Context;
-use eric_sdk::{Eric, ErrorCode};
+use eric_sdk::Eric;
 use roxmltree::Document;
 use std::{
     env::{self, current_dir},
@@ -54,9 +54,7 @@ fn test_send() {
     assert!(res.is_ok(), "{}", res.unwrap_err());
 
     let response = res.unwrap();
-    assert_eq!(response.error_code, ErrorCode::ERIC_OK as i32);
-
-    let doc = Document::parse(&response.validation_response).unwrap();
+    let doc = Document::parse(response.validation_response()).unwrap();
     let node = doc.descendants().find(|node| node.has_tag_name("Erfolg"));
     assert!(node.is_some());
 }
@@ -89,9 +87,7 @@ fn test_send_and_print() {
     assert!(res.is_ok(), "{}", res.unwrap_err());
 
     let response = res.unwrap();
-    assert_eq!(response.error_code, ErrorCode::ERIC_OK as i32);
-
-    let doc = Document::parse(&response.validation_response).unwrap();
+    let doc = Document::parse(response.validation_response()).unwrap();
     let node = doc.descendants().find(|node| node.has_tag_name("Erfolg"));
     assert!(node.is_some());
 }