+++ eric-sdk/CHANGELOG.md
@@ -12,10 +12,13 @@
   - Provide `Eric::check_xml`
   - Parse `ValidationReport` from `EricError`
   - Make `ValidationIssue` and `ValidationReport` public
+  - Add `EricApiPayload`
 - changed
   - Provide path and password as parameter in `Eric::send`
   - Make `log_path` in `Eric::new` optional, and add parameter `plugin_path`
   - Refactor logging to use `tracing` crate
+- removed
+  - Remove `error_code` from `EricResponse`
 - fixed
   - Fix dangling pointer